client/requirements.txt
Flask>=2.3.0
Flask-SocketIO>=5.3.0
together>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
gevent>=23.7.0
gevent-websocket>=0.10.1
psutil>=5.9.0
uuid 